packages/cockpitdecks_bx/pyproject.toml
# ###########################################
#
# Build
#
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
  "/cockpitdecks_bx",
  "/cockpitdecks_bx/decks/resources/types/*.yaml",
  "/cockpitdecks_bx/decks/resources/images/*.png",
]

[tool.hatch.version]
path = "cockpitdecks_bx/__init__.py"

[tool.hatch.metadata]
allow-direct-references = true


# ###########################################
#
# Project
#
[project]
name = "cockpitdecks_bx"

authors = [
  { name="Pierre M.", email="pierre@devleaks.be" },
  { name="Duane Licudi", email="dlicudi@gmail.com" }
]

description = "Cockpitdecks for Berhinger X-Touch Mini"

readme = "README.md"

license = {file = "LICENCE"}

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Games/Entertainment :: Simulation",
]

requires-python = ">=3.10"

dependencies = [
    "cockpitdecks",
    "xtouchmini @ git+https://github.com/devleaks/python-berhinger-xtouchmini.git",
    "pillow~=12.0",
]

dynamic = [
    "version"
]

[project.urls]
Homepage = "https://github.com/dlicudi/cockpitdecks"
Documentation = "https://devleaks.github.io/cockpitdecks-docs/"
Issues = "https://github.com/dlicudi/cockpitdecks/issues"
Source = "https://github.com/dlicudi/cockpitdecks"


[tool.uv.sources]
cockpitdecks = { workspace = true }

[tool.ruff]
line-length = 160
select = [
    "E",
    "F",
    "W",
]
ignore = []

[tool.ruff.format]
docstring-code-format = true

[tool.black]
line-length = 160

[tool.flake8]
max-line-length = 160

